1st stable release of Screen Saver wrote for gnome-screensaver
All satellite modules are written in Python, the screen saver in Java.
There is no C language at all.

The screen saver is now able to start and synchronize database written in rrd4j and the JVMs scanned with JMX in order to start on a past state.

The python module contain a server for ProActive user and a little proxy written in TCP/socket. The proxy listen gnome-screensaver DBus called "ActiveChanged" and send messages to the server to launch or not screensaver.jar
PAAgent is the screen saver for gnome-screensaver written in python and using GSThemeWindow, a gtk object able to get id of the screen saver window.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/agent/Common@20767 28e8926c-6b08-0410-baaa-805c5e19b8d6